requirements.txt
aiofiles==24.1.0
fastapi==0.115.11
pydantic==2.10.6
pydantic_settings==2.8.1
setuptools==76.0.0
SQLAlchemy==2.0.39
srt==3.5.3
uvicorn==0.34.0
python-multipart
psycopg2-binary==2.9.9

requests~=2.32.3
pydantic-settings~=2.8.1
starlette~=0.46.1
tweepy~=4.15.0
pytube~=15.0.0
instaloader~=4.14.1
yt-dlp~=2025.3.21
python-telegram-bot~=20.3
httpx~=0.24.1
redis~=4.5.1